Struct astral::core::collections::slot_map::SparseSlotMap

pub struct SparseSlotMap<T, Idx = u32> where
    Idx: PrimUnsignedInt + AsPrimitive<usize>,
    usize: AsPrimitive<Idx>,  { /* fields omitted */ }

A storage with stable unique keys.

See module documentation for more details.

Methods

impl<T, Idx> SlotMap<T, Idx> where
Idx: PrimUnsignedInt + AsPrimitive<usize>,
usize: AsPrimitive<Idx>,

pub fn new() -> SlotMap<T, Idx>

Construct a new, empty SparseSlotMap.

The slot map will not allocate until values are inserted.

Example

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::new();

pub fn with_capacity(capacity: usize) -> SlotMap<T, Idx>

Construct a new, empty SparseSlotMap with the specified capacity.

The slot map will be able to hold exactly capacity elements without reallocating. If capacity is 0, the vector will not allocate.

Example

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(10);

// The slot map contains no items, even though it has capacity for more
assert_eq!(map.len(), 0);

// These are all done without reallocating...
for i in 0..10 {
    map.insert(i);
}

// ...but this may make the slot map reallocate
map.insert(11);

pub fn capacity(&self) -> usize

Returns the number of elements the slot map can hold without reallocating.

Examples

use astral::core::collections::SparseSlotMap;

let map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(10);
assert_eq!(map.capacity(), 10);

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the given slot map. The collection may reserve more space to avoid frequent reallocations. After calling reserve, capacity will be greater than or equal to self.len() + additional. Does nothing if capacity is already sufficient.

Panics

Panics if the new capacity overflows usize.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(1);
map.insert(1);

map.reserve(10);
assert!(map.capacity() >= 11);

pub fn len(&self) -> Idx

Returns the number of elements in the slot map, also referred to as its 'length'.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(3);

for i in 0..3 {
    map.insert(i);
}

assert_eq!(map.len(), 3);

pub fn is_empty(&self) -> bool

Returns true if the slot map contains no elements.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(1);

assert!(map.is_empty());

map.insert(1);
assert!(!map.is_empty());

pub fn create_key(&mut self) -> Key<Idx>

Creates a new key which can be used later.

Panics

Panics if the number of elements in the slot map overflows Idx.

Example

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<u32> = SparseSlotMap::with_capacity(2);
let key1 = map.create_key();
let key2 = map.create_key();

assert!(map.is_empty());

map.insert_with_key(key2, 200);
map.insert_with_key(key1, 100);
assert_eq!(map[key1], 100);
assert_eq!(map[key2], 200);

pub fn contains_key(&self, key: Key<Idx>) -> bool

Returns if a key is stored in the map.

Example

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<u32> = SparseSlotMap::new();
let key1 = map.insert(100);
assert!(map.contains_key(key1));

A key returned from create_key() is not contained in the map until inserted with insert_with_key()

let key2 = map.create_key();
assert!(!map.contains_key(key2));
map.insert_with_key(key2, 200);
assert!(map.contains_key(key2));

pub fn insert(&mut self, value: T) -> Key<Idx>

Inserts a value into the map returning the key.

Panics

Panics if the number of elements in the slot map overflows Idx.

Example

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<u32> = SparseSlotMap::new();
let key1 = map.insert(100);
assert_eq!(map[key1], 100);
map.remove(key1);
assert!(!map.contains_key(key1));

pub fn insert_with_key(
&mut self,
key: Key<Idx>,
value: T
) -> Result<Option<T>, T>

Inserts a value at the given position. The key has to be created with create_key. It returns the previously stored value if any.

Errors

Returns back the passed value if the key is not valid.

Example

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<u32> = SparseSlotMap::new();
let key = map.create_key();

assert!(map.insert_with_key(key, 100).unwrap().is_none());
assert_eq!(map[key], 100);

If the key is used again, the value will be overwritten:

assert_eq!(map.insert_with_key(key, 200), Ok(Some(100)));
assert_eq!(map[key], 200);

If the key is not valid, the value will be passed back:

map.remove(key);
assert_eq!(map.insert_with_key(key, 300), Err(300));

pub fn remove(&mut self, key: Key<Idx>) -> Option<T>

Removes the value at the given key

Example

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<u32> = SparseSlotMap::new();
let key = map.insert(100);
assert_eq!(map.len(), 1);
map.remove(key);
assert!(map.is_empty());

Keys, which are created with create_key() can be discarded with this function as well

let key = map.create_key();
assert!(map.is_empty());
map.remove(key);
assert!(map.is_empty());

pub fn clear(&mut self)

Clears the slot map. Keeps the allocated memory for reuse.

This function must iterate over all slots, empty or not. In the face of many deleted elements it can be inefficient.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(10);

for i in 0..10 {
    map.insert(i);
}

assert_eq!(map.len(), 10);
map.clear();
assert!(map.is_empty());

pub fn retain<F>(&mut self, predicate: F) where
F: FnMut(Key<Idx>, &mut T) -> bool,

Retains only the elements specified by the predicate.

In other words, remove all key-value pairs (k, v) such that f(k, &mut v) returns false.

This function must iterate over all slots, empty or not. In the face of many deleted elements it can be inefficient.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(4);

let k1 = map.insert(1);
let k2 = map.insert(2);
let k3 = map.insert(3);
let k4 = map.insert(4);

map.retain(|key, val| key == k1 || *val % 2 == 0);
assert_eq!(map.into_iter().collect::<Vec<_>>(), vec![(k1, 1), (k2, 2), (k4, 4)]);

pub fn get(&self, key: Key<Idx>) -> Option<&T>

Returns a reference to the value corresponding to the key.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(1);

let key = map.insert(10);
assert_eq!(map.get(key), Some(&10));
map.remove(key);
assert_eq!(map.get(key), None);

pub fn get_mut(&mut self, key: Key<Idx>) -> Option<&mut T>

Returns a mutable reference to the value corresponding to the key.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<f32> = SparseSlotMap::with_capacity(1);

let key = map.insert(3.5);
if let Some(x) = map.get_mut(key) {
    *x += 3.0;
}
assert_eq!(map[key], 6.5);

pub fn iter(&self) -> Iter<T, Idx>

An iterator visiting all key-value pairs in arbitrary order. The iterator element type is (Key, &'a T).

This function must iterate over all slots, empty or not. In the face of many deleted elements it can be inefficient.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(3);

let k0 = map.insert(0);
let k1 = map.insert(1);
let k2 = map.insert(2);

let mut it = map.iter();
assert_eq!(it.next(), Some((k0, &0)));
assert_eq!(it.len(), 2);
assert_eq!(it.next(), Some((k1, &1)));
assert_eq!(it.next(), Some((k2, &2)));
assert_eq!(it.next(), None);

pub fn iter_mut(&mut self) -> IterMut<T, Idx>

An iterator visiting all key-value pairs in arbitrary order, with mutable references to the values. The iterator element type is (Key, &'a mut T).

This function must iterate over all slots, empty or not. In the face of many deleted elements it can be inefficient.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(3);

let k0 = map.insert(10);
let k1 = map.insert(20);
let k2 = map.insert(30);

for (k, v) in map.iter_mut() {
    if k != k1 {
        *v *= -1;
    }
}

assert_eq!(map.values().collect::<Vec<_>>(), vec![&-10, &20, &-30]);

pub fn keys(&self) -> Keys<T, Idx>

An iterator visiting all keys in arbitrary order. The iterator element type is Key.

This function must iterate over all slots, empty or not. In the face of many deleted elements it can be inefficient.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(3);

let k0 = map.insert(10);
let k1 = map.insert(20);
let k2 = map.insert(30);
let v: Vec<_> = map.keys().collect();
assert_eq!(v, vec![k0, k1, k2]);

pub fn values(&self) -> Values<T, Idx>

An iterator visiting all values in arbitrary order. The iterator element type is &'a T.

This function must iterate over all slots, empty or not. In the face of many deleted elements it can be inefficient.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(3);

map.insert(10);
map.insert(20);
map.insert(30);
let v: Vec<_> = map.values().collect();
assert_eq!(v, vec![&10, &20, &30]);

pub fn values_mut(&mut self) -> ValuesMut<T, Idx>

An iterator visiting all values mutably in arbitrary order. The iterator element type is &'a mut T.

This function must iterate over all slots, empty or not. In the face of many deleted elements it can be inefficient.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(3);

map.insert(10);
map.insert(20);
map.insert(30);
map.values_mut().for_each(|n| { *n *= 3 });
let v: Vec<_> = map.into_iter().map(|(_k, v)| v).collect();
assert_eq!(v, vec![30, 60, 90]);

pub fn drain(&mut self) -> Drain<T, Idx>

Creates a draining iterator that yields the removed items.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(3);

let k1 = map.insert(1);
let k2 = map.insert(2);
let k3 = map.insert(3);

let v: Vec<_> = map.drain().collect();
assert_eq!(map.len(), 0);
assert_eq!(v, vec![(k1, 1), (k2, 2), (k3, 3)]);

pub fn drain_filter<F>(&mut self, filter: F) -> DrainFilter<T, Idx, F> where
F: FnMut(Key<Idx>, &mut T) -> bool,

Clears the slot map, returning all key-value pairs as an iterator. Keeps the allocated memory for reuse.

This function must iterate over all slots, empty or not. In the face of many deleted elements it can be inefficient.

Examples

use astral::core::collections::SparseSlotMap;

let mut map: SparseSlotMap<i32> = SparseSlotMap::with_capacity(4);

let k1 = map.insert(1);
let k2 = map.insert(2);
let k3 = map.insert(3);
let k4 = map.insert(4);

let evens: Vec<_> = map.drain_filter(|_, val| *val % 2 == 0).collect();
let odds: Vec<_> = map.drain().collect();
assert!(map.is_empty());
assert_eq!(evens, vec![(k2, 2), (k4, 4)]);
assert_eq!(odds, vec![(k1, 1), (k3, 3)]);