Trait astral::core::error::ResultExt

pub trait ResultExt<T, E> {
    fn context<Kind>(self, kind: Kind) -> Result<T, Error<Kind>>;

    fn chain<Kind, Context>(
        self, 
        kind: Kind, 
        context: Context
    ) -> Result<T, Error<Kind>>
    where
        Context: Into<Box<Error + 'static + Sync + Send>>;

    fn chain_with<Kind, Context, F>(
        self, 
        kind: Kind, 
        context: F
    ) -> Result<T, Error<Kind>>
    where
        Context: Into<Box<Error + 'static + Sync + Send>>,
        F: FnOnce() -> Context;
}

Extension methods for Result.

Required Methods

fn context<Kind>(self, kind: Kind) -> Result<T, Error<Kind>>

Associates the error with an error kind.

Example

use std::{error, fmt};

use astral::core::error::ResultExt;

#[derive(Debug)]
struct CustomError;

impl fmt::Display for CustomError {
    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        fmt::Debug::fmt(self, f)
    }
}

impl error::Error for CustomError {}

#[derive(Debug, PartialEq)]
enum MyErrorKind {
    Variant,
}

let x = (|| -> Result<(), CustomError> {
    Err(CustomError)?
})().context(MyErrorKind::Variant).unwrap_err();

assert_eq!(x.kind(), &MyErrorKind::Variant);

fn chain<Kind, Context>(
self,
kind: Kind,
context: Context
) -> Result<T, Error<Kind>> where
Context: Into<Box<Error + 'static + Sync + Send>>,

Creates a new Error, associates it with an error kind and sets the old error as source.

Example

use std::{error, fmt};

use astral::core::error::ResultExt;

#[derive(Debug)]
struct CustomError;

impl fmt::Display for CustomError {
    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        fmt::Debug::fmt(self, f)
    }
}

impl error::Error for CustomError {}

#[derive(Debug)]
enum MyErrorKind {
    Variant,
}

impl fmt::Display for MyErrorKind {
    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        fmt::Debug::fmt(self, f)
    }
}

let x = (|| -> Result<(), CustomError> {
    Err(CustomError)?
})().chain(MyErrorKind::Variant, "An error occured").unwrap_err();

assert_eq!(x.to_string(), "An error occured");

fn chain_with<Kind, Context, F>(
self,
kind: Kind,
context: F
) -> Result<T, Error<Kind>> where
Context: Into<Box<Error + 'static + Sync + Send>>,
F: FnOnce() -> Context,

Creates a new Error, associates it with an error kind and sets the old error as source by applying the provided closure FnOnce() -> impl Into<Box<dyn error::Error + Send + Sync>>.

Example

use std::{error, fmt};

use astral::core::error::ResultExt;

#[derive(Debug)]
struct CustomError;

impl fmt::Display for CustomError {
    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        fmt::Debug::fmt(self, f)
    }
}

impl error::Error for CustomError {}

#[derive(Debug)]
enum MyErrorKind {
    Variant,
}

impl fmt::Display for MyErrorKind {
    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        fmt::Debug::fmt(self, f)
    }
}

let x = (|| -> Result<(), CustomError> {
    Err(CustomError)?
})().chain_with(MyErrorKind::Variant, || "An error occured").unwrap_err();

assert_eq!(x.to_string(), "An error occured");