Module astral::core::log

Structures and macros for logging.

The logging module provides the logging API that abstracts over the actual logging implementation.

If no logging implementation is selected, the facade falls back to a "noop" implementation that ignores all log messages. The overhead in this case is very small - just an integer load, comparison and jump.

A log request consists of a target, a level, and a body. A target is a string which defaults to the module path of the location of the log request, though that default may be overridden. Logger implementations typically use the target to filter requests based on some user configuration.

For more information see the log crate.

Use

The basic use of the log crate is through the five logging macros: error!, warn!, info!, debug! and trace! where error! represents the highest-priority log messages and trace! the lowest. The log messages are filtered by configuring the log level to exclude messages with a lower priority. Each of these macros accept format strings similarly to println!.

Example

use astral::core::log::{info, warn};

pub fn shave_the_yak(yak: &mut Yak) {
    info!(target: "yak_events", "Commencing yak shaving for {:?}", yak);

    loop {
        match find_a_razor() {
            Ok(razor) => {
                info!("Razor located: {}", razor);
                yak.shave(razor);
                break;
            }
            Err(err) => {
                warn!("Unable to locate a razor: {}, retrying", err);
            }
        }
    }
}

Macros

debug
Logs a message at the debug level.

error
Logs a message at the error level.

info
Logs a message at the info level.

log
The standard logging macro.

log_enabled
Determines if a message logged at the specified level in that module will be logged.

trace
Logs a message at the trace level.

warn
Logs a message at the warn level.

Structs

Metadata
Metadata about a log message.

MetadataBuilder
Builder for Metadata.

ParseLevelError
The type returned by from_str when the string doesn't match any of the log levels.

Record
The "payload" of a log message.

RecordBuilder
Builder for Record.

SetLoggerError
The type returned by set_logger if set_logger has already been called.

TerminalLogger
Provides a stderr/out based Logger implementation.

Enums

Level
An enum representing the available verbosity levels of the logger.

LevelFilter
An enum representing the available verbosity level filters of the logger.

Constants

STATIC_MAX_LEVEL
The statically resolved maximum log level.

Traits

Log
A trait encapsulating the operations required of a logger.

Functions

logger
Returns a reference to the logger.

max_level
Returns the current maximum log level.

set_boxed_logger
Sets the global logger to a Box<Log>.

set_logger
Sets the global logger to a &'static Log.

set_max_level
Sets the global maximum log level.