Module astral::core::string

Structures for holding strings.

This module contains the Name and the Text type. While both can hold strings, Name is optimized for strings with a numeric suffix. Texts implement Deref<Target=str>, which is not the case for Name, because of the optimization.

Examples

There are multiple ways to create a new Text or a new Name from a string literal:

use astral::core::string::{Text, Name};

let t = Text::from("foo");
let n: Name = "foo".into();
assert_eq!(t, n);

A Text can be converted into &'static str:

use astral::core::string::Text;

let n = Text::from("foo");
let s: &'static str = n.as_str();

assert_eq!("foo", s)

If you have a slice of valid UTF-8 bytes, you can make a Text or a Name out of it.

use astral::core::string::Text;

let sparkle_heart = &[240, 159, 146, 150];

// We know these bytes are valid, so we'll use `unwrap()`.
let sparkle_heart = Text::from_utf8(sparkle_heart).unwrap();

assert_eq!("💖", sparkle_heart);

let bytes = sparkle_heart.as_bytes();

assert_eq!(bytes, [240, 159, 146, 150]);

Structs

Name
A UTF-8 encoded, immutable string optimized for numeric suffixes.

Text
A UTF-8 encoded, immutable string.

Utf8Error
Errors which can occur when attempting to interpret a sequence of u8 as a string.

Utf16Error
A possible error value when converting a Name or Text from a UTF-16 byte slice.

Constants

MAX_STRINGS
The maximum number of unique strings like Text or Name.

MAX_STRING_LENGTH
The maximum length of one string like Text or Name.