Struct astral::resource::assets::Catalog

pub struct Catalog<'loc> { /* fields omitted */ }

A collection of Namespaces.

Each Namespace is associated with a NamespaceId for accessing the Namespace at a later point.

A VirtualFileSystem can be used directly with a Location.

Methods

impl<'loc> Catalog<'loc>

pub fn new() -> Catalog<'loc>

Construct a new empty Catalog.

Example

use astral::resource::assets::Catalog;

let catalog = Catalog::new();
assert!(catalog.is_empty());

pub fn with_capacity(capacity: usize) -> Catalog<'loc>

Construct a new, empty Catalog with the specified capacity.

The Catalog will be able to hold exactly capacity Namespaces without reallocating. If capacity is 0, the Catalog will not allocate.

Example

use astral::resource::assets::{Catalog, Namespace};

let mut catalog = Catalog::with_capacity(1);

assert_eq!(catalog.len(), 0);

// This can be done without reallocating...
catalog.add_namespace(Namespace::new());

// ...but this may make the catalog reallocate
catalog.add_namespace(Namespace::new());

pub fn len(&self) -> usize

Returns the number of Namespaces in the Catalog.

Example

use astral::resource::assets::{Catalog, Namespace};

let mut catalog = Catalog::new();

assert_eq!(catalog.len(), 0);
catalog.add_namespace(Namespace::new());
assert_eq!(catalog.len(), 1);

pub fn is_empty(&self) -> bool

Returns true if the Catalog contains no Namespaces.

Example

use astral::resource::assets::{Catalog, Namespace};

let mut catalog = Catalog::new();

assert!(catalog.is_empty());
catalog.add_namespace(Namespace::new());
assert!(!catalog.is_empty());

pub fn add_namespace(&mut self, namespace: Namespace<'loc>) -> NamespaceId

Adds a new Namespace and returns its NamespaceId to query it at a later time.

Example

use astral::resource::assets::{Catalog, Namespace};

let mut catalog = Catalog::new();

assert_eq!(catalog.len(), 0);
let namespace_id = catalog.add_namespace(Namespace::new());
assert_eq!(catalog.len(), 1);

pub fn get_namespace(
&self,
namespace_id: NamespaceId
) -> Option<&Namespace<'loc>>

Returns the Namespace as reference of the given NamespaceId if any or None.

Example

use astral::resource::assets::{Catalog, Namespace};

let mut catalog = Catalog::new();

let namespace_id = catalog.add_namespace(Namespace::new());
assert!(catalog.get_namespace(namespace_id).is_some());

pub fn get_namespace_mut(
&mut self,
namespace_id: NamespaceId
) -> Option<&mut Namespace<'loc>>

Returns the Namespace as mutable reference of the given NamespaceId if any or None.

Example

use astral::resource::assets::{Catalog, Namespace};

let mut catalog = Catalog::new();

let namespace_id = catalog.add_namespace(Namespace::new());
assert!(catalog.get_namespace_mut(namespace_id).unwrap().is_empty());

pub fn remove_namespace(
&mut self,
namespacae_id: NamespaceId
) -> Option<Namespace<'loc>>

Removes the Namespace of the given NamespaceId and returns it.

Example

use astral::resource::assets::{Catalog, Namespace};

let mut catalog = Catalog::new();

let namespace_id = catalog.add_namespace(Namespace::new());
assert!(catalog.remove_namespace(namespace_id).is_some());

pub fn contains_namespace(&self, namespace_id: NamespaceId) -> bool

Returns true if the Namespace of the given NamespaceId exists.

Example

use astral::resource::assets::{Catalog, Namespace};

let mut catalog = Catalog::new();

let namespace_id = catalog.add_namespace(Namespace::new());
assert!(catalog.contains_namespace(namespace_id));

pub fn iter_namespaces(
&self
) -> impl Iterator<Item = (NamespaceId, &Namespace<'loc>)>

Returns an Iterator over all NamespaceIds and Namespaces in the Catalog.

Example

use astral::resource::assets::{Catalog, Namespace};

let mut catalog = Catalog::new();

catalog.add_namespace(Namespace::new());
for (id, namespace) in catalog.iter_namespaces() {
    // do something with the namespace
}

pub fn create(
&mut self,
location: Location,
virtual_file_system_index: Option<VirtualFileSystemIndex>
) -> Option<Result<impl Write, Error<ErrorKind>>>

Opens a file in write-only mode at the given VirtualFileSystem.

See Namespace::create for more infos.

Example

use astral::core::string::Name;
use astral::resource::assets::{Catalog, FileSystem, Location, Namespace};

let mut namespace = Namespace::new();
let cwd_index = namespace.add_virtual_file_system(FileSystem::new(".", false)?)?;

let mut catalog = Catalog::new();
let namespace_id = catalog.add_namespace(namespace);

let location = Location::new(namespace_id, Name::from("file.txt"));
catalog.create(location, Some(cwd_index));

pub fn create_new(
&mut self,
location: Location,
virtual_file_system_index: Option<VirtualFileSystemIndex>
) -> Option<Result<impl Write, Error<ErrorKind>>>

Opens a file in write-only mode at the given VirtualFileSystem.

See Namespace::create_new for more infos.

Example

use astral::core::string::Name;
use astral::resource::assets::{Catalog, FileSystem, Location, Namespace};

let mut namespace = Namespace::new();
let cwd_index = namespace.add_virtual_file_system(FileSystem::new(".", false)?)?;

let mut catalog = Catalog::new();
let namespace_id = catalog.add_namespace(namespace);

let location = Location::new(namespace_id, Name::from("file.txt"));
catalog.create_new(location, Some(cwd_index));

pub fn exists(&self, location: Location) -> bool

Returns whether the Catalog is aware of the file and the entity exists.

See Namespace::exists for more infos.

Example

use astral::core::string::Name;
use astral::resource::assets::{Catalog, FileSystem, Location, Namespace};

let mut namespace = Namespace::new();

// fill namespace

let mut catalog = Catalog::new();
let namespace_id = catalog.add_namespace(namespace);

let location = Location::new(namespace_id, Name::from("does_not_exists.txt"));

assert_eq!(catalog.exists(location), false);

pub fn modified(
&self,
location: Location
) -> Option<Result<SystemTime, Error<ErrorKind>>>

Returns the last modification time of the file if the Catalog is aware of it.

See Namespace::modified for more infos.

Example

use astral::core::string::Name;
use astral::resource::assets::{Catalog, FileSystem, Location, Namespace};

let mut namespace = Namespace::new();

// fill namespace

let mut catalog = Catalog::new();
let namespace_id = catalog.add_namespace(namespace);

let location = Location::new(namespace_id, Name::from("file.txt"));

println!("{:?}", catalog.modified(location));

pub fn open(
&self,
location: Location
) -> Option<Result<impl Read, Error<ErrorKind>>>

Opens the file in read-only mode. Returns None, if the Catalog is not aware of it.

See Namespace::open for more infos.

Example

use astral::core::string::Name;
use astral::resource::assets::{Catalog, FileSystem, Location, Namespace};

let mut namespace = Namespace::new();
let cwd_index = namespace.add_virtual_file_system(FileSystem::new(".", false)?)?;

let mut catalog = Catalog::new();
let namespace_id = catalog.add_namespace(namespace);

let location = Location::new(namespace_id, Name::from("file.txt"));
if let Some(read) = catalog.open(location) {
    let file = read?;
}

pub fn remove(
&mut self,
location: Location
) -> Option<Result<(), Error<ErrorKind>>>

Remove the file. Returns Some(Result<()>), if the Catalog is aware of the file. Result determines if the removal was successful. Returns None otherwise

See Namespace::remove for more infos.

Example

use astral::core::string::Name;
use astral::resource::assets::{Catalog, FileSystem, Location, Namespace};

let mut namespace = Namespace::new();
let cwd_index = namespace.add_virtual_file_system(FileSystem::new(".", false)?)?;

let mut catalog = Catalog::new();
let namespace_id = catalog.add_namespace(namespace);

let location = Location::new(namespace_id, Name::from("file.txt"));
if let Some(result) = catalog.remove(location) {
    println!("removing file: {:?}", result);
}