Struct astral::resource::assets::FileSystem

pub struct FileSystem<'str, 'ass, H = BuildHasherDefault<Murmur3>> where
    H: 'str,  { /* fields omitted */ }

A FileSystem is a view into the systems file system.

Methods

impl<'str, 'ass, H> FileSystem<'str, 'ass, H>

pub fn new<P>(
root: P,
asset_subsystem: &'ass Subsystem,
string_subsystem: &'str Subsystem<H>
) -> Result<FileSystem<'str, 'ass, H>, Error<ErrorKind>> where
P: Into<PathBuf>,

Construct a new FileSystem at the specified root path. If wished, the file system can search files in a recursive manner.

Example

use astral::{
    core::string::Name,
    resource::assets::{FileSystem, VirtualFileSystem},
};

let file_system = FileSystem::new(".", &asset_subsystem, &string_subsystem)?;
let file_name = Name::new("a.txt", &string_subsystem);
let file = file_system.open(file_name.into())?;

Important traits for FileSystem<'str, '_, H>
impl<'str, '_, H> VirtualFileSystem<'str, H> for FileSystem<'str, '_, H> where
H: BuildHasher + Debug + Send + Sync,

pub fn test() -> FileSystem<'str, 'ass, H>

test func