Struct astral::resource::assets::Namespace

pub struct Namespace<'loc> { /* fields omitted */ }

A Namespace contains multiple, various VirtualFileSystems.

The files inside of a VirtualFileSystem are cached for a faster access. The cache can be recreated with reload.

Methods

impl<'loc> Namespace<'loc>

pub fn new() -> Namespace<'loc>

Construct a new empty Namespace.

Example

use astral::resource::assets::Namespace;

let namespace = Namespace::new();
assert!(namespace.is_empty());

pub fn with_capacity(
virtual_file_systems: usize,
files: usize
) -> Namespace<'loc>

Construct a new, empty Namespace with the specified capacity.

The Namespace will be able to hold exactly virtual_file_systems elements file systems and files unique files without reallocating. If both parameters are 0, the Namespace will not allocate.

Example

use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::with_capacity(1, 100);

// The namespace contains no items, even though it has capacity for more
assert_eq!(namespace.virtual_file_systems(), 0);

// This can be done without reallocating,
// assuming there are no more than 100 files...
namespace.add_virtual_file_system(FileSystem::new(".", false)?);

// ...but this may make the namespace reallocate
namespace.add_virtual_file_system(FileSystem::new("..", false)?);

pub fn virtual_file_systems(&self) -> usize

Returns the number of VirtualFileSystems in the Namespace.

Example

use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::with_capacity(1, 100);

assert_eq!(namespace.virtual_file_systems(), 0);
namespace.add_virtual_file_system(FileSystem::new(".", false)?);
assert_eq!(namespace.virtual_file_systems(), 1);

pub fn files(&self) -> usize

Returns the number of files in the Namespace.

Example

use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::new();

assert_eq!(namespace.files(), 0);
namespace.add_virtual_file_system(FileSystem::new(".", false)?);
assert!(namespace.files() > 0);

pub fn is_empty(&self) -> bool

Returns true if the Namespace contains no VirtualFileSystems or no files.

Example

use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::new();

assert!(namespace.is_empty());
namespace.add_virtual_file_system(FileSystem::new(".", false)?);
assert!(!namespace.is_empty());

pub fn add_virtual_file_system<impl Into>>(
&mut self,
virtual_file_system: impl Into>
) -> Result<VirtualFileSystemIndex, Error<ErrorKind>> where
impl Into>: Into<Box<VirtualFileSystem + 'loc>>,

Adds a new VirtualFileSystem and returns its VirtualFileSystemIndex to query it at a later time.

Example

use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::new();
let cwd_index = namespace.add_virtual_file_system(FileSystem::new(".", false)?)?;

pub fn remove_virtual_file_system(
&mut self,
virtual_file_system_index: VirtualFileSystemIndex
) -> Option<Box<VirtualFileSystem + 'loc>>

Removes a VirtualFileSystem by its index, which where returned by add_virtual_file_system. Returns the file system if any.

Example

use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::new();
assert!(namespace.is_empty());
let cwd_index = namespace.add_virtual_file_system(FileSystem::new(".", false)?)?;
assert!(!namespace.is_empty());
let vfs = namespace.remove_virtual_file_system(cwd_index);
assert!(namespace.is_empty());
assert!(vfs.is_some());

pub fn clear(&mut self)

Removes all VirtualFileSystems from the Namespace.

Example

use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::new();

namespace.add_virtual_file_system(FileSystem::new(".", false)?);
namespace.add_virtual_file_system(FileSystem::new("..", false)?);
assert_eq!(namespace.virtual_file_systems(), 2);
namespace.clear();
assert!(namespace.is_empty());

pub fn reload(
&mut self,
virtual_file_system_index: VirtualFileSystemIndex
) -> Result<(), Error<ErrorKind>>

Reloads the VirtualFileSystem at the given index and updates the internal cache. This may take some time.

Example

use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::new();

let cwd_index = namespace.add_virtual_file_system(FileSystem::new(".", false)?)?;
namespace.reload(cwd_index);

pub fn iter(
&self
) -> impl Iterator<Item = (&VirtualFileSystem, Name)>

Returns an Iterator over all VirtualFileSystems in the Namespace.

Example

use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::new();

namespace.add_virtual_file_system(FileSystem::new(".", false)?);
for vfs in namespace.iter() {
    // do something with the file systems
}

pub fn create(
&mut self,
name: Name,
virtual_file_system_index: Option<VirtualFileSystemIndex>
) -> Option<Result<impl Write, Error<ErrorKind>>>

Opens a file in write-only mode at the given VirtualFileSystem.

If no VirtualFileSystemIndex is provided, the first VirtualFileSystem will be used, which is not read-only.

This function will create a file if it does not exist, and will truncate it if it does.

Example

use astral::core::string::Name;
use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::new();

let cwd_index = namespace.add_virtual_file_system(FileSystem::new(".", false)?)?;
namespace.create(Name::from("a.txt"), Some(cwd_index));

pub fn create_new(
&mut self,
name: Name,
virtual_file_system_index: Option<VirtualFileSystemIndex>
) -> Option<Result<impl Write, Error<ErrorKind>>>

Creates a file in write-only mode at the given VirtualFileSystem.

If no VirtualFileSystemIndex is provided, the first VirtualFileSystem will be used, which is not read-only.

No file is allowed to exist at the target location, also no (dangling) symlink.

Example

use astral::core::string::Name;
use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::new();

let cwd_index = namespace.add_virtual_file_system(FileSystem::new(".", false)?)?;
namespace.create_new(Name::from("a.txt"), Some(cwd_index));

pub fn exists(&self, name: Name) -> bool

Returns whether the Namespace is aware of the file and the entity exists.

Example

use astral::core::string::Name;
use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::new();

let cwd_index = namespace.add_virtual_file_system(FileSystem::new(".", false)?)?;
assert_eq!(namespace.exists(Name::from("does_not_exist.txt")), false);

pub fn modified(
&self,
name: Name
) -> Option<Result<SystemTime, Error<ErrorKind>>>

Returns the last modification time of the file if the Namespace is aware of it.

Example

use astral::core::string::Name;
use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::new();

let cwd_index = namespace.add_virtual_file_system(FileSystem::new(".", false)?)?;
println!("{:?}", namespace.modified(Name::from("file.txt")));

pub fn open(
&self,
name: Name
) -> Option<Result<impl Read, Error<ErrorKind>>>

Opens the file in read-only mode. Returns None, if the Namespace is not aware of it.

Example

use astral::core::string::Name;
use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::new();

let cwd_index = namespace.add_virtual_file_system(FileSystem::new(".", false)?)?;
if let Some(read) = namespace.open(Name::from("file.txt")) {
    let file = read?;
}

pub fn remove(&mut self, name: Name) -> Option<Result<(), Error<ErrorKind>>>

Remove the file. Returns Some(Result<()>), if the Namespace is aware of the file. Result determines if the removal was successful. Returns None otherwise

Example

use astral::core::string::Name;
use astral::resource::assets::{Namespace, FileSystem};

let mut namespace = Namespace::new();

let cwd_index = namespace.add_virtual_file_system(FileSystem::new(".", false)?)?;
if let Some(result) = namespace.remove(Name::from("file.txt")) {
    println!("removing file: {:?}", result);
}