Trait astral::resource::assets::VirtualFileSystem

pub trait VirtualFileSystem<'str, H = BuildHasherDefault<Murmur3>>: Send + Sync + Debug {
    fn name(&self) -> Name<'str, H>;

    fn readonly(&self) -> bool;

    fn iter(
        &'a self
    ) -> Result<Box<dyn Iterator<Item = Name<'str, H>> + 'a>, Error<ErrorKind>>;

    fn create(
        &mut self, 
        path: Name<'str, H>
    ) -> Result<Box<dyn Write + 'static>, Error<ErrorKind>>;

    fn create_new(
        &mut self, 
        path: Name<'str, H>
    ) -> Result<Box<dyn Write + 'static>, Error<ErrorKind>>;

    fn exists(&self, path: Name<'str, H>) -> bool;

    fn modified(
        &self, 
        path: Name<'str, H>
    ) -> Result<SystemTime, Error<ErrorKind>>;

    fn open(
        &self, 
        path: Name<'str, H>
    ) -> Result<Box<dyn Read + 'static>, Error<ErrorKind>>;

    fn remove(&mut self, path: Name<'str, H>) -> Result<(), Error<ErrorKind>>;
}

A virtual file system is an abstraction to a concrete file system with which you can read, write and create files.

The most primitive file system is the file system of the operating system. Therefore an implementation is given with FileSystem.

Required methods

fn name(&self) -> Name<'str, H>

Returns the Name of the file system.

fn readonly(&self) -> bool

Returns if the file system is read-only.

fn iter(
&'a self
) -> Result<Box<dyn Iterator<Item = Name<'str, H>> + 'a>, Error<ErrorKind>>

Returns an Iterator over all files in the file system.

fn create(
&mut self,
path: Name<'str, H>
) -> Result<Box<dyn Write + 'static>, Error<ErrorKind>>

Opens a file in write-only mode.

This function will create a file if it does not exist, and will truncate it if it does.

fn create_new(
&mut self,
path: Name<'str, H>
) -> Result<Box<dyn Write + 'static>, Error<ErrorKind>>

Creates a file in write-only mode.

No file is allowed to exist at the target location, also no (dangling) symlink.

fn exists(&self, path: Name<'str, H>) -> bool

Returns whether the path points at an existing entity.

fn modified(&self, path: Name<'str, H>) -> Result<SystemTime, Error<ErrorKind>>

Returns the last modification time at this entity.

fn open(
&self,
path: Name<'str, H>
) -> Result<Box<dyn Read + 'static>, Error<ErrorKind>>

Attempts to open a file in read-only mode.

fn remove(&mut self, path: Name<'str, H>) -> Result<(), Error<ErrorKind>>

Removes a file from the filesystem.

Loading content...